18/1/21 Rearrange the flow of -checkFile. That is, func of checkDCCVTAConstraint()

+++ src/utility.cpp
@@ -144,7 +144,7 @@ bool isRealNumber(string str)
 // All combinations are stored in "comblist"
 //
 /////////////////////////////////////////////////////////////////////
-void combination(long loc, int total, int catchn, vector<long> gencomb, vector<vector<long> > *comblist)
+void combination(long loc, long int total, int catchn, vector<long> gencomb, vector<vector<long> > *comblist)
 {
 	gencomb.push_back(loc-1);
 	for( long loop = loc;loop <= (total-catchn) && (catchn > 0);loop++ )
